+++ lib/exirc/client.ex
@@ -8,6 +8,8 @@ defmodule ExIrc.Client do
   alias ExIrc.Channels, as: Channels
   alias ExIrc.Utils,    as: Utils
 
+  alias ExIrc.Client.Transport, as: Transport
+
   # Client internal state
   defmodule ClientState do
     defstruct event_handlers:   [],
@@ -263,34 +265,6 @@ defmodule ExIrc.Client do
      login_time:        state.login_time,
      debug?:            state.debug?,
      event_handlers:    state.event_handlers]
-  end
-
-
-  ###############
-  # Transport
-  ###############
-
-  defmodule Transport do
-    def connect(%{ssl?: false}, host, port, options) do
-      :gen_tcp.connect(host, port, options)
-    end
-    def connect(%{ssl?: true}, host, port, options) do
-      :ssl.connect(host, port, options)
-    end
-
-    def send(%{ssl?: false, socket: socket}, data) do
-      :gen_tcp.send(socket, data)
-    end
-    def send(%{ssl?: true, socket: socket}, data) do
-      :ssl.send(socket, data)
-    end
-
-    def close(%{ssl?: false, socket: socket}) do
-      :gen_tcp.close(socket)
-    end
-    def close(%{ssl?: true, socket: socket}) do
-      :ssl.close(socket)
-    end
   end
 
   ###############
